+++ radical_translations/conftest.py
@@ -1,8 +1,12 @@
 import pytest
+from django.core import management
 from django.test import RequestFactory
 
 from radical_translations.users.models import User
 from radical_translations.users.tests.factories import UserFactory
+
+
+pytestmark = pytest.mark.django_db
 
 
 @pytest.fixture(autouse=True)
@@ -18,3 +22,9 @@ def user() -> User:
 @pytest.fixture
 def request_factory() -> RequestFactory:
     return RequestFactory()
+
+
+@pytest.fixture
+@pytest.mark.django_db
+def vocabulary():
+    management.call_command("vocab", "init", verbosity=0)