+++ HISTORY.rst
@@ -3,7 +3,13 @@
 History
 =======
 
+LATEST_VERSION
+---
+
+* Add command to import Events from GSX
+* Add command to import Organisations from GSX
+
 0.1.0 (2020-02-05)
-------------------
+---
 
 * Initial data models